(Portage version: 2.2.0_alpha10/hg/Linux x86_64, unsigned Manifest commit)

+++ net-misc/flexget/flexget-1.0_beta1845.ebuild
@@ -11,14 +11,22 @@ DISTUTILS_SRC_TEST="setup.py"
 
 inherit distutils
 
-MY_P="FlexGet-${PV/_beta/r}"
+if [[ ${PV} == 9999 ]]; then
+	inherit subversion
+	ESVN_REPO_URI="http://svn.flexget.com/trunk"
+	KEYWORDS=""
+else
+	MY_P="FlexGet-${PV/_beta/r}"
+	SRC_URI="http://download.flexget.com/unstable/${MY_P}.tar.gz"
+	KEYWORDS="~amd64"
+	S=${WORKDIR}/${MY_P}
+fi
+
 DESCRIPTION="A multipurpose automation tool for content like torrents, nzbs, podcasts, comics, etc."
 HOMEPAGE="http://flexget.com/"
-SRC_URI="http://download.flexget.com/unstable/${MY_P}.tar.gz"
 
 LICENSE="MIT"
 SLOT="0"
-KEYWORDS="~amd64"
 IUSE="test"
 
 RDEPEND="dev-python/setuptools
@@ -37,14 +45,17 @@ DEPEND="${RDEPEND}
 	dev-python/paver
 	test? ( >=dev-python/nose-0.11 )"
 
-S=${WORKDIR}/${MY_P}
-
 src_prepare() {
 	# Prevent setup from grabbing nose from pypi
 	sed -e /setup_requires/d -i pavement.py || die
 
-	# Use system paver for building
-	rm -f paver-minilib.zip
+	if [[ ${PV} == 9999 ]]; then
+		# Generate setup.py
+		paver generate_setup
+	else
+		# Remove bundled paver
+		rm -f paver-minilib.zip
+	fi
 
 	distutils_src_prepare
 }
